refactor(primitives,tx-graph): move DepositInfo to tx-graph and mark old one as deprecated
