+++ crates/primitives/Cargo.toml
@@ -28,6 +28,3 @@ serde.workspace = true
 sha2.workspace = true
 strata-p2p-types.workspace = true
 thiserror.workspace = true
-
-[dev-dependencies]
-serde_json.workspace = true
