+++ crates/primitives/src/deposit.rs
@@ -23,6 +23,7 @@ use crate::{
 
 /// The deposit information  required to create the Deposit Transaction.
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[deprecated = "this has been moved to strata_bridge-tx-graph as `DepositRequestData` and will be removed along with the `strata-bridge-agent` crate in the future"]
 pub struct DepositInfo {
     /// The deposit request transaction outpoints from the users.
     deposit_request_outpoint: OutPoint,
@@ -236,322 +237,5 @@ impl DepositInfo {
         let unsigned_tx = create_tx(tx_ins, tx_outs);
 
         Ok(unsigned_tx)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::fs;
-
-    use bitcoin::{script::Instruction, Network};
-    use secp256k1::PublicKey;
-    use strata_primitives::{operator::OperatorPubkeys, params::OperatorConfig};
-
-    use super::*;
-    use crate::{
-        build_context::TxBuildContext,
-        errors::{BridgeTxBuilderError, DepositTransactionError},
-        test_utils::{
-            create_drt_taproot_output, generate_keypairs, generate_pubkey_table,
-            generate_xonly_pubkey,
-        },
-    };
-
-    /// Loads the sidesystem params from the test file.
-    ///
-    /// If pubkeys are supplied, it updates the operator config in the params with the provided
-    /// pubkeys as `wallet_pk`.
-    fn test_sidesystem_params<Pks>(pubkeys: Option<Pks>) -> RollupParams
-    where
-        Pks: IntoIterator<Item = PublicKey>,
-    {
-        let test_rollup_params = fs::read_to_string("../../test-data/rollup_params.json")
-            .expect("could not read test rollup params");
-
-        let mut params = serde_json::from_str::<RollupParams>(&test_rollup_params)
-            .expect("rollup-params in test-data must have valid structure");
-
-        if let Some(pubkeys) = pubkeys {
-            params.operator_config = OperatorConfig::Static(
-                pubkeys
-                    .into_iter()
-                    .map(|wallet_pk| {
-                        let wallet_pk = wallet_pk.x_only_public_key().0;
-                        OperatorPubkeys::new([2u8; 32].into(), wallet_pk.serialize().into())
-                    })
-                    .collect(),
-            );
-        }
-
-        params
-    }
-
-    #[test]
-    fn test_create_spend_infos() {
-        let (operator_pubkeys, _) = generate_keypairs(10);
-        let operator_pubkeys = generate_pubkey_table(&operator_pubkeys);
-
-        let deposit_request_outpoint = OutPoint::null();
-        let recovery_xonly_pk = generate_xonly_pubkey();
-
-        let refund_delay = 1008;
-        let (drt_output_address, _take_back_leaf_hash) =
-            create_drt_taproot_output(operator_pubkeys.clone(), recovery_xonly_pk, refund_delay);
-        let self_index = 0;
-
-        let tx_builder = TxBuildContext::new(Network::Regtest, operator_pubkeys, self_index);
-        let deposit_amt = Amount::from_int_btc(1);
-
-        // Correct merkle proof
-        let deposit_info = DepositInfo::new(
-            deposit_request_outpoint,
-            1,
-            [0u8; 20].to_vec(),
-            deposit_amt,
-            recovery_xonly_pk,
-            drt_output_address.address().script_pubkey(),
-        );
-
-        let result = deposit_info.validate(&tx_builder, refund_delay);
-        assert!(
-            result.is_ok(),
-            "should build the prevout for DT from the deposit info, error: {:?}",
-            result.err().unwrap()
-        );
-
-        // Handles incorrect merkle proof
-        let random_xonly_pubkey = generate_xonly_pubkey();
-        let deposit_info = DepositInfo::new(
-            deposit_request_outpoint,
-            1,
-            [0u8; 20].to_vec(),
-            deposit_amt,
-            random_xonly_pubkey,
-            drt_output_address.address().script_pubkey(),
-        );
-
-        let result = deposit_info.validate(&tx_builder, refund_delay);
-
-        assert!(result.as_ref().err().is_some());
-        assert!(
-            matches!(
-                result.unwrap_err(),
-                BridgeTxBuilderError::DepositTransaction(
-                    DepositTransactionError::InvalidTapLeafHash
-                ),
-            ),
-            "should handle the case where the supplied merkle proof is wrong"
-        );
-    }
-
-    #[test]
-    fn test_create_unsigned_tx() {
-        let (operator_pubkeys, _) = generate_keypairs(10);
-        let operator_pubkeys = generate_pubkey_table(&operator_pubkeys);
-
-        let deposit_request_outpoint = OutPoint::null();
-        let recovery_xonly_pk = generate_xonly_pubkey();
-
-        let refund_delay = 1_008;
-        let (drt_output_address, _take_back_leaf_hash) =
-            create_drt_taproot_output(operator_pubkeys.clone(), recovery_xonly_pk, refund_delay);
-        let self_index = 0;
-
-        let tx_builder = TxBuildContext::new(Network::Regtest, operator_pubkeys, self_index);
-        let deposit_amt = Amount::from_int_btc(1);
-
-        let deposit_info = DepositInfo::new(
-            deposit_request_outpoint,
-            1,
-            [0u8; 20].to_vec(),
-            deposit_amt,
-            recovery_xonly_pk,
-            drt_output_address.address().script_pubkey(),
-        );
-
-        let tag = b"alpen";
-        let deposit_amt = Amount::from_int_btc(1);
-        let result =
-            deposit_info.create_unsigned_tx(&tx_builder, deposit_amt, tag, 20, refund_delay);
-        assert!(
-            result.is_ok(),
-            "should build the prevout for DT from the deposit info, error: {:?}",
-            result.err().unwrap()
-        );
-
-        let unsigned_tx = result.unwrap();
-        assert_eq!(unsigned_tx.input.len(), 1);
-        assert_eq!(unsigned_tx.output.len(), 2);
-    }
-
-    #[test]
-    fn test_construct_psbt() {
-        let (operator_pubkeys, _) = generate_keypairs(10);
-        let operator_pubkeys = generate_pubkey_table(&operator_pubkeys);
-
-        let deposit_request_outpoint = OutPoint::null();
-        let recovery_xonly_pk = generate_xonly_pubkey();
-
-        let refund_delay = 1008;
-        let (drt_output_address, _take_back_leaf_hash) =
-            create_drt_taproot_output(operator_pubkeys.clone(), recovery_xonly_pk, refund_delay);
-        let self_index = 0;
-
-        let tx_builder = TxBuildContext::new(Network::Regtest, operator_pubkeys, self_index);
-        let deposit_amt = Amount::from_int_btc(1);
-
-        let deposit_info = DepositInfo::new(
-            deposit_request_outpoint,
-            1,
-            [0u8; 20].to_vec(),
-            deposit_amt,
-            recovery_xonly_pk,
-            drt_output_address.address().script_pubkey(),
-        );
-
-        let deposit_amt = Amount::from_int_btc(1);
-        let result = deposit_info.construct_psbt(
-            &tx_builder,
-            &PegOutGraphParams::default(),
-            &test_sidesystem_params::<Vec<_>>(None),
-        );
-        assert!(
-            result.is_ok(),
-            "should build the prevout for DT from the deposit info, error: {:?}",
-            result.err().unwrap()
-        );
-
-        let psbt = result.unwrap();
-        assert_eq!(psbt.unsigned_tx.input.len(), 1);
-        assert_eq!(psbt.unsigned_tx.output.len(), 2);
-
-        // test with invalid EL address
-        const INVALID_LENGTH: usize = 21;
-        let deposit_info = DepositInfo::new(
-            deposit_request_outpoint,
-            1,
-            [0u8; 21].to_vec(),
-            deposit_amt,
-            recovery_xonly_pk,
-            drt_output_address.address().script_pubkey(),
-        );
-
-        let result = deposit_info.construct_psbt(
-            &tx_builder,
-            &PegOutGraphParams::default(),
-            &test_sidesystem_params::<Vec<_>>(None),
-        );
-        assert!(
-            result.is_err_and(|e| matches!(
-                e,
-                BridgeTxBuilderError::DepositTransaction(
-                    DepositTransactionError::InvalidEeAddressSize(INVALID_LENGTH, 20)
-                )
-            )),
-            "should handle the case where the EL address is invalid"
-        );
-
-        // test with invalid x-only pk
-        let random_xonly_pk = generate_xonly_pubkey();
-
-        let deposit_info = DepositInfo::new(
-            deposit_request_outpoint,
-            1,
-            [0u8; 20].to_vec(),
-            deposit_amt,
-            random_xonly_pk,
-            drt_output_address.address().script_pubkey(),
-        );
-
-        let result = deposit_info.construct_psbt(
-            &tx_builder,
-            &PegOutGraphParams::default(),
-            &test_sidesystem_params::<Vec<_>>(None),
-        );
-        assert!(
-            result.is_err_and(|e| matches!(
-                e,
-                BridgeTxBuilderError::DepositTransaction(
-                    DepositTransactionError::InvalidTapLeafHash
-                )
-            )),
-            "should handle the case where the supplied merkle proof is wrong"
-        );
-    }
-
-    #[test]
-    fn test_deposit_tx_metadata() {
-        let network = Network::Regtest;
-
-        let (operator_pubkeys, _) = generate_keypairs(5);
-        let operator_pubkeys = generate_pubkey_table(&operator_pubkeys);
-        let sidesystem_params = test_sidesystem_params(Some(operator_pubkeys.0.values().copied()));
-
-        let tx_build_context = TxBuildContext::new(network, operator_pubkeys, 0);
-
-        let recovery_xonly_pubkey = generate_xonly_pubkey();
-        let pegout_graph_params = PegOutGraphParams::default();
-
-        let take_back_script =
-            drt_take_back(recovery_xonly_pubkey, pegout_graph_params.refund_delay);
-        let take_back_script_hash =
-            TapNodeHash::from_script(&take_back_script, LeafVersion::TapScript);
-
-        let spend_path = SpendPath::Both {
-            internal_key: tx_build_context.aggregated_pubkey(),
-            scripts: &[take_back_script],
-        };
-
-        let (deposit_request_addr, _) = create_taproot_addr(&network, spend_path)
-            .expect("must be able to generate taproot address for drt");
-
-        let stake_index = 1;
-        let ee_address = [0u8; 20];
-        let total_amount = Amount::from_int_btc(11);
-        let deposit_info = DepositInfo::new(
-            OutPoint::null(),
-            stake_index,
-            ee_address.to_vec(),
-            total_amount,
-            recovery_xonly_pubkey,
-            deposit_request_addr.script_pubkey(),
-        );
-
-        let psbt = deposit_info
-            .construct_psbt(&tx_build_context, &pegout_graph_params, &sidesystem_params)
-            .expect("must be able to construct signing data");
-        let deposit_tx = psbt.unsigned_tx;
-
-        let expected_metadata = [
-            pegout_graph_params.tag.as_bytes(),
-            &stake_index.to_be_bytes(),
-            &ee_address,
-            take_back_script_hash.as_ref(),
-            &total_amount.to_sat().to_be_bytes(),
-        ]
-        .concat();
-
-        let Some(op_return_out) = deposit_tx.output.get(1) else {
-            panic!("must have a second output");
-        };
-        let script_pubkey = &op_return_out.script_pubkey;
-
-        assert!(
-            script_pubkey.is_op_return(),
-            "second output must be an OP_RETURN"
-        );
-
-        let mut instructions = script_pubkey.instructions();
-        instructions.next(); // consume the OP_RETURN instruction
-
-        let Some(Ok(Instruction::PushBytes(data))) = instructions.next() else {
-            panic!("the second output must have some PushBytes instruction and data");
-        };
-
-        assert_eq!(
-            data.as_bytes(),
-            expected_metadata,
-            "the metadata in the second output must be equal to the expected metadata"
-        );
     }
 }
