+++ crates/test-utils/src/lib.rs
@@ -7,6 +7,7 @@
 pub mod arbitrary_generator;
 pub mod bitcoin;
 pub mod bitcoin_rpc;
+pub mod deposit;
 pub mod musig2;
 pub mod prelude;
 pub mod tx;
